Import the tag parsing tests from cachemachine
There were almost no tests of the tag parsing logic and this (for
now until we create a shared model library) will be the canonical
location for this code, so import the tests from cachemachine and
reformat them a little to be less verbose.

Put this test at the top level in anticipation of tag parsing
moving out of models into an abstract data type, since that feels
like a better way of thinking about it.

Moves models/tag.py coverage from 77% to 89%.