Refactor creation of build component of tags
trailing_parts_to_semver_build_component had a bug ([^\w|\.]
includes | since [] is not ()) and the structure was a bit hard to
read.  Rename it to make it clear that it's an internal helper
function and not part of the API, add a proper docstring, and
restructure the code to make the logic a bit more obvious.

Convert the tag parsing helper functions to class methods from
static methods, and call them from cls instead of the hard-coded
name of the class (otherwise, subclassing and overriding internal
methods will not work correctly).