Plumb Docker storage tests farther through
Rather than generating a username and password inside the test,
use the mechanism for fetching Docker credentials to use the default
credentials from the test configuration.

This uncovered a minor bug in the Docker storage client where it was
adding a spurious extra slash to the URL before the repository.  Fix
the repository and path properties to avoid this.

Start using Factory properly in the test suite to generate the
Docker storage object.