Clean up configuration handling
Rework how the runtime section of the Configuration object is
handled to avoid needing to set empty defaults and then override
them later.  One parameter is initialized from an environment
variable; for that, use BaseSettings and configure the environment
variable to use, since Pydantic can handle that directly.  One
parameter gets a default from a disk file; do that with a factory
function and an environment variable override.  The final variable
wants the path to the loaded configuration file, which we can patch
into the parsed YAML before initializing the object.

Convert the path to the configuration file to a Path object
everywhere, with corresponding changes to the config dependency and
its callers.