Overhaul Docker secret configuration
Make the Docker storage client support multiple registries and move
the registry and repository parameters to arguments to its methods.
This is cleaner conceptually and will also be required if we ever
want to pre-pull more than one image in the future.

Refactor DockerCredentialsMap into a new DockerCredentialStore that
supports read and write operations (although not update, yet).  Move
some of the code that is simple transformations of the data into the
DockerCredentials class following the principles about where to put
code surrounding models.

Move construction of the tag map into the prepuller and out of the
storage layer since this is business logic, not a low-level operation
against the storage layer.  This was why the storage layer had to know
about the recommended tag, which it should never need to know about.

These changes plus moving the path to the Docker secrets file into the
main configuration instead of putting special-case logic into the
Factory class provided an opportunity to eliminate the runtime portion
of the configuration entirely.