Add set method to DockerCredentialStore
Finish the API for DockerCredentialStore so that it can be used to
manipulate a credential store.  (This package doesn't use this, but
we will probably lift this code into Safir and use it in Phalanx for
tools for updating pull secrets.)

Remove the host information from the DockerCredentials object.
It was never used and makes storing and loading more tedious.