Overhaul tag parsing
Move StandaloneRSPTag to a new RSPImageTag class in a separate
domain model that provides an abstract data type. The rest of
the current models/tag.py code will move in a separate change.
Rework the parsing code to hopefully be easier to follow and a
bit more succinct.

Comparing incomparable tag types now behaves like comparing any
other incomparable types.  The cost is a less specific and verbose
error message, but given that this indicates a programming error,
that seems fine for the additional simplicity.