Refactor the prepuller
Extract the prepuller logic for determining remote and cached
images and calculating the prepull targets, menus, and prepull
status into a new ImageService class. Rewrite it on top of the
new RSPImageTag and RSPImage abstract data types and their
collections, which can now do most of the filtering work. The
new image service is now used in handlers where previously the
PrepullerAribtrator was used.

This new image service only asks for digests for images that are
prepull targets, so has less impact on the Docker registry API.
However, it doesn't lock in that decision; we can decide to
retrieve hashes for all images with some minor additional work.
The public API for available images drops the human-readable
descriptions for alias tags (I think they add more clutter than
useful information) and distinguishes between the main tag and
the aliases using the new RSPImage data structure.

Ensuring that every pod will be launched with a reference with
a digest will be added in a subsequent commit.

The image service now uses a single delay for polling both Docker
and Kubernetes (which now may be a bit long, but it can be updated
in the future with experience).

What remains of the prepuller is only the logic to ask the image
service for images to prepull and start tasks to do so. Simplify
its coordination with the image service by using asyncio.Event
instead of polling and timeouts, so the prepuller is kicked off
each time the image status has been refreshed. The prepuller
creates one background task per node with missing images and pulls
those images in the same sort order that is used in the menu (with
the minor variation that all alias tags are pulled first in
alphabetical order rather than forcing the recommended tag first).

The resulting prepuller is small enough that it can be collapsed
into a single class.