Mock Gafaelfawr with respx
Replace the patched mock Gafaelfawr storage class with respx
mocking so the actual GafaelfawrStorageClient can be used. This
required some changes to the test data format to include the
tokens that map to particular users.

The get_scopes method was never called, so remove it and thus the
need to mock the token-info endpoint.