Add end-to-end handler test
Add a test for the user-status endpoint that goes through the
FastAPI layer. This turned up a bunch of problems with the mock of
the Kubernetes storage object mostly due to incomplete mocking.
Change strategies and use a Mock with a spec instead of a
handwritten class, with special attention for the methods that
Mock cannot handle dynamically.