Avoid unnecessary Gafaelfawr calls
For all of the handlers that only need to know the authenticated
username, get that from the HTTP request headers rather than making
a call to Gafaelfawr.

Fix the route registration for the list of users with running labs
to avoid adding a spurious slash, similar to the previous fix for
user-info.