Convert text dumps of objects to JSON
Whenever possible, compare objects in pytest by converting them to
dictionaries and comparing the dictionaries with assert. pytest will
then show a rich diff if there are differences, rather than showing
both strings as unhelpfully long lines.

Convert the text dumps of objects to JSON, and compare the JSON for
this improved error reporting. Convert DisplayImages to a Pydantic
model from a dataclass to make generating a dict easier.