Compare Kubernetes objects for pod tests
Unfortunately, the serialization of Kubernetes models includes
every possible setting and is possibly unstable in the future if
APIs ever change (to add new mount types, for instance). Comparing
to a disk serialization is therefore not very maintainable or
understandable. Instead, hard-code the Kubernetes objects into the
test and then compare dictionaries.