Include build information when sorting tags
semver ignores the build component of the version when comparing
versions, but that's where we put the cycle and cycle build
information. We therefore don't want to ignore it because we want
tags with a newer cycle or cycle build to sort after ones with an
older cycle or cycle build for otherwise identical versions, and
tags with a cycle to sort after ones without a cycle. This ensures
that when presenting tags in reverse sorted order, ones with a
newer cycle or cycle build will be shown first.