Add RSPImage abstract type and its collection
Continuing to lay the groundwork for a clearer expression of the
complex image analysis algorithm, this introduces abstract data
types for tracking images. An image is a tag annotated with its
registry, repository, and digest, and is used to drive prepulling
and to compare against when determining what images are cached.

As with RSPImageTagCollection, the operations on RSPImageCollection
are chosen to satisfy the requirements of subsequent changes.
This is where nearly all logic for alias handling, including
resolving aliases to their underlying tags, is done.