Ensure all labs are spawned with a digest
The decision to not retrieve digests for all available images means
that labs spawned from the dropdown menu may not have digests.
However, we always want to record the digest of the image when
spawning the pod for debugging purposes and better user display.
The lab service also had some left-over reference and tag parsing
code to reconstruct information about the image from the reference
passed in via the lab creation form.

Correct this by adding a new method to the image service to create
an RSPImage from a Docker reference, and then use the RSPImage as
the source of data to encode in the pod environment ConfigMap. In
the process, change image to reference in the lab options model and
refactor reference parsing so that it can be shared between the lab
service and cached image analysis.