Switch to full Kubernetes API mocking
Stop using the mock Kubernetes storage object in favor of a full
mock of the Kubernetes API for the methods that the lab controller
calls. This is temporarily implemented as a subclass of the mock
class provided by Safir, but those changes will be rolled back into
Safir once they have been thoroughly exercised.

This uncovered a timing problem with one lab service API test, where
deletion was handled by a background task and therefore the user map
didn't update immediately. Handle that with a delay that should be
more than long enough for the background deletion to clean things up.

Remove some code from the Kubernetes storage layer that is handled by
the Watch API. Setting _preload_content isn't an optimization; it
tells the API to return the raw HTTP response rather than the
deserialized object, and is always set internally by the Watch API
(and should never be set otherwise). There is similarly no need to
provide resource_version unless we're resuming a previous watch, since
Watch handles setting it automatically.