Compare Kubernetes objects using serialization
Rather than giving up on serializing Kubernetes objects, use a
smarter serialization that omits all of the None parameters.
Drop the separate test for the build_volumes method of the lab
manager because all of its output is incorporated into the pod
spec and therefore can be tested by that existing test.