Fix start/stop logic for background tasks
For the image service and prepuller, make the background task
logic consistent and stop reusing a scheduler. It looks like it
may be unsafe to reuse one after it's been closed, so create the
scheduler on start and use its existence as the flag to determine
if background tasks are running.