Fix typing issues caught by mypy 1.1.1
Enable the pydantic mypy plugin for better type checking and fix
legitimate (or at least reasonable) typing issues caught by mypy
1.1.1. Most of these are using explicit enum objects rather than
their string values as initializers for enum fields, but there were
also a few type mismatches and incomplete Optional tagging or
checking for None.

This reduces the mypy errors with 1.1.1 down to only the errors
about missing initializers for optional fields, which is a bug in
mypy's interaction with the pydantic plugin and should be fixed in
the next pydantic release.