Fix handling of image selection in lab UserOptions
Treat empty lists in UserOptions input equivalent to None. Add a
test for UserOptions to ensure that DROPDOWN_SENTINEL_VALUE is
handled correctly, lists are converted to regular elements, and
booleans are converted correctly.

If both image_list and image_dropdown are set (but only those two),
use image_list and ignore image_dropdown. Due to how our form is
designed, image_dropdown will always be set, but if image_list is
not the sentinel value, it should override.

Stop configuring UserOptions with use_enum_values. This was an
attempt to put strings into the dict() output for easier testing,
but the test suite has been refactored to not need this, and this
put the string directly into the model attribute as well, breaking
other parts of the code.