Clean up headers sent with lab controller requests
Rework how the headers are constructed when sending requests to
be a bit more obvious about whether the request is authenticated
with the user token or the admin token. Stop sending Content-Type
headers with requests with no body; this is arguably an HTTP
protocol error. Use the admin token for the poll route, since I
believe it needs it, rather than trying to use the user's token.