Add resource settings to lab containers
In order to be allowed to set a ResourceQuota for the user's
namespace, every pod container has to also set resources so that
Kubernetes can validate that they're not exceeding their quota.
Add the code to do that, setting the lab container quotas based on
the chosen lab size, since the intent for the namespace quota is
to potentially be larger than the lab quota and allow for spawning
of additional pods.

Set the same quota for the init container, even though it's unlikely
to use that many resources, since that's simple and has no effect
on the overall requirements for the pod.

Refactor SizeManager a bit to use actual methods instead of
disguising them as properties and to generate only the resources
requested, rather than generating all of them and then throwing
most of them away.