Test full Kubernetes object creation
Rather than testing individual options on the lab manager to create
bits and pieces of the user's environment, use our mock Kubernetes
to retrieve all resources that were created by the test and check all
of them against a JSON file of expected results.

This unfortunately makes the files stored in ConfigMaps a bit more
annoying to review, but it seems worth it.

Add api_version and kind to all objects as we create them. This isn't
necessary (Kubernetes figures it out), but it makes reviewing the test
data easier.