Simplify handlers and Safir configuration
Use path_prefix as the setting name for the prefix under which to
mount all routes, matching times-square. Mount all handler routers
at that prefix and move the rest of the URL into the handler
definition, since that's easier to understand when reading the
handler code.

Remove a bunch of unneeded docstrings, following the pattern that
handlers should use parameters to the route registration rather than
docstrings for documentation. This simplifies the logic for what goes
into the OpenAPI specification.

Rename indexes.py to index.py so that the handler names match the
test names, and to match other Safir applications.

Stop supporting a configurable logger name. We dropped support for
that in the template a while back, since we never used it in practice.

Support standard environment variables for the Safir settings for
consistency with other applications.

Remove response models that were never used and make some of the
descriptions shorter since the generated documentation is a bit
space-constrained.