Test form generation through a handler
We may as well test form generation by going through the front door
of the application instead of bypassing it and calling directly into
the service. This also allows deleting the direct test of the image
service method for generating menu images, since those results are
incorporated into the user form.