Add test for watching events during delayed spawn
Test that we can watch events while a pod is spawning and will see
the messages from events in the namespace. This uncovered various
problems with the mock Kubernetes API.

Add some more debuging to the event reflector that was helpful in
figuring out where the test suite wasn't working.