Refactor event management
The previous way of handling events used a queue, which meant that
only one reader was supported and if anyone else requested the pod
events for a user, one or both of the listeners would miss events
and possibly hang forever.

Rework the event manager to return separate iterators for every
caller that use an asyncio Event to tell them when there are new
events to read. Improve the event models a bit in the process.