Add support for Google Artifact Registry
Subclass ImageService for Docker and GAR and start teasing it apart
into the generic functions and the ones that are specific to a given
source of images. Add a Google Artifact Registry storage backend and
test mock and a test case for using it to sync rather than Docker.

Add some new functions to retrieve test data. The plan is for this
approach to eventually replace the current fixtures.

Fix a bug in how image references were generated in the non-prepulled
section of the image list uncovered by the new tests.