Add subsetting and more alias tracking to RSPImage
For Google Artifact Registry support, we'll need to support subset
on RSPImageCollection as well as RSPImageTagCollection. Add that
and a test for it.

Improve alias tracking for images in an RSPImageCollection,
particularly in case of two non-alias, non-unknown tags that alias
each other and the case of multiple aliases for the same tag. Use a
new aliased flag to determine whether an image is the target of an
alias tag rather than the existence of something in the aliases set,
since the latter may be because two "real" images (w_2077_43 and
w_2077_44, for instance) alias each other because the same image was
reused.