Refactor image source handling
Rather than subclassing ImageService, introduce a new intermediate
service, ImageSource, with Docker and GAR specializations, that sits
between the image service and the underlying storage client. This
provides a cleaner separation of data between the layers of services
and hopefully more straightforward code.

Redo how image sources are configured to use a shared source key
with a mandatory type field that is set to either docker or google.
This works better with Pydantic and makes the typing less complex.
Pass this down to the Docker storage layer instead of separate
registry and repository options.

Improve the way requested tags are mapped to images so that we
resolve alias tags to their underlying images more often.