Refactor recording of image presence on nodes
Use the RSPImageCollection method to record when an image is seen
in a node cache, rather than manipulating the images directly.
This chases aliases properly.

While doing this, I noticed a race condition where node information
for images would be wrong between pulling new remote images and
pulling cache data from the Kubernetes cluster during an image
service update cycle. This was previously handled by the image
service, but when the data was moved to a source, the synchronization
was lost. Solve that problem alongside the node update by adding a
method to the image sources that updates node information for all
images and sets the newly-retrieved image information as active.

Also do node updates in mark_prepulled, which means that the test
output for the GAR prepuller configuration is now completely correct.