Add tests for menu generation when using GAR
With Google Artifact Repository, we're now properly tracking aliases
across both prepulled and non-prepulled images. That originally
broke assumptions about menu generation, since now images are hidden
if they have any aliases, even if those aliases aren't in the main
part of the menu and are only in the dropdown. (This problem wouldn't
happen with Docker because with Docker we don't track aliases for
images that aren't prepulled.) Test that this problem was fixed by
the RSPImageCollection changes.

We don't suppress as aliased images that have an alias tag pointing
to their digest, but which are not the primary target. This only
happens when there are multiple non-alias tags with the same digest.
It's arguable whether to suppress these as well, but since the alias
tag takes its description from the primary target, having that image
go missing when it might be an expected older weekly image (that
happens to have the same digest) seemed confusing. We can always
revisit this later if it comes up.

Add a utility function to write test data as JSON. This is not used
normally by the test suite but is handy when writing new tests.