Simplify Kubernetes object creation
Now that the mock copies the standard Kubernetes API behavior of
setting the kind, api_version, and metadata.namespace based on the
parameters to the API functions, remove all of that from the objects
we create. This allows simplifying how standard metadata is created
for user lab objects.

Remove some trailing commas and adjust some variable names for a
bit of additional formatting compactness and (hopefully) clarity.