Add node tracking to RSPImage
Rather than manipulating the node information for an RSPImage
directly in the image service, add a method to RSPImageCollection
to mark an image as seen on a node. This allows RSPImageCollection
to handle the alias chasing correctly so that all tags that alias
that tag also get accurate node cache information.