Rework how aliased tag detection is handled
The previous approach to aliased tag handling (a non-alias tag that
is the target of an alias tag) didn't take into account the
possibility that the alias tag wasn't present in the collection.
This happens in the Google Artifact Registry case when constructing
the menu, since the alias tag may not be included in the menu and
therefore we don't want the underlying image tag to be considered
aliased for image construction purposes.

Get rid of the not-granular-enough aliased flag on an image and
instead determine whether a tag is aliased by checking the collection
to see if an alias tag pointing to it is present.