Improve error handling for bad Docker references
Turn the InvalidDockerReferenceError into a rich exception that
understands how to format itself as a structured JSON error, and
annotate it with the location of the error. Add an exception
handler for all exceptions descended from ValidationError that
automatically generates structured errors.

This particular error is not very important, but this establishes
the groundwork for doing this for other errors that should return
structured errors and can be handled by an exception handler.