Fix cycle handling with Google Artifact Registry
Cycle filtering wasn't implemented on RSPImageCollection or
plumbed through to the Google Artifact Registry storage layer.
Do that, and add a test suite for cycle handling with Google
Artifact Registry as well.

This uncovered a few more capitalization problems when an unknown
tag with a cycle is later resolved into an alias. Fix those by
always title-casing the display name when resolving the alias.