Add structured exception for Kubernetes errors
Flesh out KubernetesError with the ability to report itself to
Slack and the ability to create itself from an ApiException and
capture details of that exception. Ensure all Kubernetes
ApiException errors are converted to it. Be more restrictive in
what Kubernetes errors we catch in the storage layer, since other
unexpected errors should start as unexpected until we understand
them.

Simplify the error handling for user state reconciliation to only
skip past 404 errors and to abort the first time there's a fatal
Kubernetes API exception. We will want to add retry logic at a
higher level, but continuing without pause when Kubernetes is
raising API exceptions is likely to be pointless.