Rework exception handling layer for lab spawning
Move all Kubernetes object creation into a background thread,
rather than starting only with the lab creation, so that all error
messages are reported via the events API and none have to be
reported as a direct response from the route handler.

Standardize the exception handling during lab creation and deletion
to happen at the top level and uniformly report failure events. Fix
the event message on success events for deleting pods.

Add a stringification of KubernetesError and use the string value
of the exception, without additional prefixes, as the error message
to the user.

Fix the logic when waiting for pod creation if the lab hasn't been
created yet.

Remove the now-obsolete services/lab_test test, since everything it
tests is now tested through the handler.