Use asyncio locking for progress events
Rather than using a delay similar to kubespawner to wait for new
events, use the same system that the Nublado lab controller uses
internally: a separate asyncio.Event for each progress call, each
of which are set when each new event arrives.

Don't bother to clean up the events when the progress call completes.
The next spawn for that user, or the disposal of the spawner object,
will take care of that, and the spawner already stores other state
for the user until its reuse for another spawn, so this shouldn't be
significant additional memory consumption.