Exclude in-progres spawns from reconciliation
Now that we're reconciling state in the background, we don't want
to reconcile state for labs that are currently in the process of
being spawned. Doing that may cause us to race with ourselves and
incorrectly update state.

Add new in-progress tracking to ensure that we don't do that. In
so doing, move more of the state tracking of initial lab creation
into LabStateManager, allowing some simplification of the API.

Change the progress percentage of the first event to 1% (there
didn't seem to be a good reason to start at 2%).

Remove checking of internal data structures so that we raise an
uncaught exception if we mess up the state tracking. Eventually
we'll do something more sophisticated for Redis, but in the meantime
don't hide logic errors.