Refactor pod spawn task management
Previously, when someone requested that a lab be spawned, this
created a background task recorded in the LabManager that created
the Kubernetes objects and then waited for the pod to finish
starting. The drawback of this approach is that nothing ever reaped
this background task to property collect its return status or
exception.

Now that LabStateManager exists, hand over management of the spawner
task to it. Provide it with a callback that creates the Kubernetes
objects, and move waiting for the pod to start into LabStateManager.
These threads are now reaped by a long-running reaper task. Move the
exception handling and reporting for the spawner task into
LabStateManager as well.

In addition to the benefit of better task tracking and management,
this also lays the groundwork for spawning a monitoring task during
startup if we detect a lab spawn in progress. It also allows making
more of the event posting and state update machinery internal to
LabStateManager, since it now knows when lab spawns complete and
doesn't have to be told by other code.