Collect all lab data together
Rather than maintaining multiple parallel data structures that
could potentially get out of date, collect all of the internal data
about a user's lab into an internal dataclass and maintain that in
a single dictionary.

As a side effect, this means we no longer retain events for a lab
after it's deleted (but events can still be monitored while lab
deletion is in progress). After consideration, I think this is
correct REST semantics, and JupyterHub shouldn't need to retrieve
events for a successful deletion.