Merge waiting for pod with event monitoring
Merge the method to wait for a pod to start with the method that
monitors events for that pod and eliminate the extra background
task for event injection. Use that same method for the prepuller.

Add a new configurable spawn timeout instead of hard-coding it,
and pass that timeout into the Kubernetes watch, using a shorter
watch for the network connection timeout.