Merge waiting for pod with event monitoring
Merge the method to wait for a pod to start with the method that
monitors events for that pod and eliminate the extra background
task for event injection. Use that same method for the prepuller.

Add a new configurable spawn timeout instead of hard-coding it,
and pass that timeout into the Kubernetes watch, using a shorter
watch for the network connection timeout.

This removes the suppression of duplicate messages, since it's a
bit more awkward to do in this format. We can add it back in, but
I wanted to try running this version first and see to what degree
duplicate messages are an issue.