Redo how labs are deleted
Use the same division of labor between LabManager and LabStateManager
as for lab creation, which lets us eliminate more public APIs from
LabStateManager for the individual events involved in deletion. As
with lab spawning, LabStateManager now handles the exceptions,
although lab deletion is done in the foreground since JupyterHub does
it in the foreground anyway and therefore no purpose is served by
adding background deletion.