Fix URL of running lab
The Jupyter lab takes its base URL from JUPYTERHUB_SERVICE_PREFIX
in the lab environment, which is always set by JupyterHub. Although
there is a redirect from the top-level URL, ensure that the internal
URL we return is fully correct so that JupyterHub doesn't complain
about it and change it during startup.

Make JUPYTERHUB_SERVICE_PREFIX required in the lab environment so
that we don't have to do error handling, reporting, or fallback.
This is unconditionally sent by JupyterHub, so we can make it part
of the API contract.