Fix lab state reconciliation
When the lab controller was restarted, its internal data structures
were only partly refreshed from the running labs in Kubernetes. A lot
of data was missing or incorrect, such as all of the environment
variables. This ran afoul of the new validator to ensure that
JUPYTERHUB_SERVICE_PREFIX is set.

Fully restore internal state by also reading the ConfigMap and
ResourceQuota objects, and add the data that isn't otherwise easily
discernable from lab resources (full name and group name information,
which only otherwise exist in the templated NSS files) to annotations
on the pod.

This introduces a new LabBuilder object, which eventually will hold all
the code to construct Kubernetes objects for a user's lab environment
(moved from LabManager, which will then only be about managing lab
lifecycle). For the time being, this only holds the methods that need
to be shared between LabManager and LabStateManager.