Handle pending labs found in reconciliation
If during lab reconciliation we discover a lab in pending status,
we need to monitor it so that we see when it starts, just as if it
were a lab we created. Otherwise, labs that were pending when the
lab controller restarts will stay pending forever and never show as
ready.