Improve Kubernetes exception handling
Add a new exception, MissingObjectError, to be used instead of
MissingSecretError. Unlike the exception it replaces, it contains
more metadata about the missing object and is Slack-reportable.

Remove the body of a Kubernetes error from the summary line of the
Slack message. It's added later in the Slack message, so this was
both ugly and redundant.