Go back to polling for pod completion
Unfortunately, Kubernetes doesn't send a namespace event when a pod
changes phase to running, so we can't rely entirely on events
triggering polls of the pod phase. This requires unfortunately much
more complex logic, since we have to start a background task to watch
namespace events and ensure that it's shut down properly when the pod
is running.

This currently uses polling, which creates some unfortunate delays.
A future iteration should use a watch instead, but that will require
more extensive work on the mock.