Request all events when waiting for pod to start
The Kubernetes mock provided by Safir now matches the behavior
of Kubernetes of only showing events after the current time when
a watch is started without specifying a resource version. For now,
specify a resource version of 0 when watching for pod events, since
the event indicating the pod has started will be published before
we start the watch. This will be fixed in a later change once the
pod watch becomes more sophisticated.