Add support for persistent volume claims
USDF requires creation of PVCs for each Jupyter lab pods to mount
the underlying storage for home space and project space. Add support
for PVCs as a source for mounted lab volumes and, in that case,
create a PVC with the given parameters. The supported configuration
parameters are a subset of the PVC model that is sufficient for the
current USDF requirements.

While making this change, refactor the configuration syntax for lab
volumes to separate the types of supported volumes into different
Pydantic models so that Pydantic can do validation for us.