Decrease grace period for lab shutdown
Follow kubespawner and kill the lab with a grace period of 1s before
deleting the namespace. The lab apparently doesn't do anything useful
with its grace period, and this will take the user out of the lab
faster and avoid the pop-up error messages from the lab being removed
from its service.