Log REST spawner exceptions before translating
JupyterHub appears to swallow the original backtrace of an exception
raised with a from clause in the spawner exception translation
wrapper. Explicitly log the original exception before translating it
in an attempt to get better backtraces into the logs.

Since the wrapper is outside of the class, it doesn't have ready
access to the regular logger, so just use Python logging.